refactor: extract AppHeader, AppNavbar, add NavbarContext
Split header and navbar into separate components (AppHeader, AppNavbar) and introduce NavbarContext for managing navbar state. Updated Default layout to use new components and context provider, improving modularity and maintainability.
